Update Events
Auto Hide Events Past Date
