+++ app/_components/events.tsx
@@ -1,55 +1,5 @@
 import FadeInTransition from "./scroll-transition";
-
-const events = [
-  {
-    id: 1,
-    title: "99 Rising Tabling",
-    date: "2023-09-27",
-    startTime: "09:00",
-    endTime: "14:00",
-    location: "HSSB Courtyard",
-  },
-  {
-    id: 2,
-    title: "Fall Kickoff General Meeting",
-    date: "2023-10-09",
-    startTime: "20:00",
-    endTime: "22:00",
-    location: "Broida 1610",
-  },
-  {
-    id: 3,
-    title: "Big Little Pickup: Virtual Speed Dating",
-    date: "2023-10-11",
-    startTime: "20:00",
-    endTime: "22:00",
-    location: "Zoom",
-  },
-  {
-    id: 4,
-    title: "Big Little Pickup: In Person Speed Dating",
-    date: "2023-10-13",
-    startTime: "20:00",
-    endTime: "22:00",
-    location: "Lot 22 Top Flor",
-  },
-  {
-    id: 5,
-    title: "Big Little Pickup: Beach Day",
-    date: "2023-10-21",
-    startTime: "14:00",
-    endTime: "16:00",
-    location: "Goleta Beach",
-  },
-  {
-    id: 6,
-    title: "Big Little Pickup: Arts and Crafts Night",
-    date: "2023-10-25",
-    startTime: "20:00",
-    endTime: "22:00",
-    location: "SRB Multipurpose Room",
-  },
-];
+import events from "../_data/eventData";
 
 const Events = () => {
   const formatTime = (time: string) => {
@@ -71,6 +21,10 @@ const Events = () => {
     return formattedDate;
   };
 
+  const currentDate = new Date();
+  const futureDate = new Date();
+  futureDate.setDate(currentDate.getDate() - 1);
+
   return (
     <div className="bg-gray-100 py-8">
       <div className="container mx-auto px-4 sm:px-6 lg:px-8">
@@ -79,21 +33,23 @@ const Events = () => {
             Upcoming Events
           </h2>
           <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-            {events.map((event) => (
-              <div key={event.id} className="bg-white rounded-lg shadow p-4">
-                <h3 className="text-xl font-bold mb-2">{event.title}</h3>
-                <p className="text-gray-600 mb-2">
-                  <strong>Date:</strong> {formatDate(event.date)}
-                </p>
-                <p className="text-gray-600 mb-2">
-                  <strong>Time:</strong> {formatTime(event.startTime)} -{" "}
-                  {formatTime(event.endTime)} PST
-                </p>
-                <p className="text-gray-600">
-                  <strong>Location:</strong> {event.location}
-                </p>
-              </div>
-            ))}
+            {events
+              .filter((event) => new Date(event.date) >= futureDate) // Filter events with dates in the future
+              .map((event) => (
+                <div key={event.id} className="bg-white rounded-lg shadow p-4">
+                  <h3 className="text-xl font-bold mb-2">{event.title}</h3>
+                  <p className="text-gray-600 mb-2">
+                    <strong>Date:</strong> {formatDate(event.date)}
+                  </p>
+                  <p className="text-gray-600 mb-2">
+                    <strong>Time:</strong> {formatTime(event.startTime)} -{" "}
+                    {formatTime(event.endTime)} PST
+                  </p>
+                  <p className="text-gray-600">
+                    <strong>Location:</strong> {event.location}
+                  </p>
+                </div>
+              ))}
           </div>
         </FadeInTransition>
       </div>
