+++ app/_data/bigLittleEvents.ts
@@ -1,0 +1,34 @@
+const events = [
+    {
+      name: "Virtual Speed Dating",
+      description: "Join us for Virtual Speed Dating!",
+      date: "2023-10-11",
+      location: "Zoom",
+    },
+    {
+      name: "In Person Speed Dating",
+      description: "Come out to Lot 22 to meet your potential big/little!",
+      date: "2023-10-13",
+      location: "Lot 22 top floor",
+    },
+    {
+      name: "Boba Mixer",
+      description: "Join us for a night of boba and fun!",
+      date: "2023-10-18",
+      location: "Theater and Dance 1701",
+    },
+    {
+      name: "Beach Day",
+      description: "Come out to the beach with us for a day of sun and fun!",
+      date: "2023-10-21",
+      location: "Goleta Beach",
+    },
+    {
+      name: "Art and Game Night",
+      description:
+        "Enjoy a night of arts and crafts and games! Last chance to meet your potential big/little!",
+      date: "2023-10-25",
+      location: "TBD",
+    },
+  ];
+export default events;
