+++ app/_data/eventData.ts
@@ -1,0 +1,52 @@
+const events = [
+  {
+    id: 1,
+    title: "99 Rising Tabling",
+    date: "2023-9-29",
+    startTime: "09:00",
+    endTime: "14:00",
+    location: "HSSB Courtyard",
+  },
+  {
+    id: 2,
+    title: "Fall Kickoff General Meeting",
+    date: "2023-10-09",
+    startTime: "20:00",
+    endTime: "22:00",
+    location: "Broida 1610",
+  },
+  {
+    id: 3,
+    title: "Big Little Pickup: Virtual Speed Dating",
+    date: "2023-10-11",
+    startTime: "20:00",
+    endTime: "22:00",
+    location: "Zoom",
+  },
+  {
+    id: 4,
+    title: "Big Little Pickup: In Person Speed Dating",
+    date: "2023-10-13",
+    startTime: "20:00",
+    endTime: "22:00",
+    location: "Lot 22 Top Flor",
+  },
+  {
+    id: 5,
+    title: "Big Little Pickup: Beach Day",
+    date: "2023-10-21",
+    startTime: "14:00",
+    endTime: "16:00",
+    location: "Goleta Beach",
+  },
+  {
+    id: 6,
+    title: "Big Little Pickup: Arts and Crafts Night",
+    date: "2023-10-25",
+    startTime: "20:00",
+    endTime: "22:00",
+    location: "SRB Multipurpose Room",
+  },
+];
+
+export default events;
